fix(freetype): 修复函数返回值类型错误导致的潜在问题
- 将 `sizeMetricsGetYPPEM` 返回类型从 `long` 修改为 `int`，更符合实际使用需求
- 同步更新相关头文件声明以避免类型不一致

+++ include/freetype_cinterop.h
@@ -124,7 +124,7 @@ int sizeMetricsGetXPPEM(long sizeMetrics);
 
 long sizeMetricsGetXScale(long sizeMetrics);
 
-long sizeMetricsGetYPPEM(long sizeMetrics);
+int sizeMetricsGetYPPEM(long sizeMetrics);
 
 long sizeMetricsGetYScale(long sizeMetrics);
